# nest-next-openapi/next/openapi.json
{"openapi":"3.0.0","paths":{"/auth/otp/send":{"post":{"description":"Send a one-time password to the user's email","operationId":"AuthController_sendOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}},"responses":{"200":{"description":"OTP sent successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}},"400":{"description":"Invalid email format","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}}},"summary":"Send OTP","tags":["auth"]}},"/auth/otp/verify":{"post":{"description":"Verify the one-time password and authenticate the user","operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}},"responses":{"200":{"description":"OTP verified successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}},"400":{"description":"Invalid OTP code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}}},"summary":"Verify OTP","tags":["auth"]}},"/auth/logout":{"post":{"description":"Logout the current user and clear authentication cookies","operationId":"AuthController_logout","parameters":[],"responses":{"204":{"description":"Logout successful"},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}}},"summary":"Logout","tags":["auth"]}},"/auth/token/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/me":{"get":{"description":"Get information about the currently authenticated user","operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"User information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaHolderClass"}}}}},"summary":"Get current user","tags":["auth"]}},"/auth/google":{"get":{"operationId":"AuthController_googleInit","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/google/callback":{"get":{"operationId":"AuthController_googleCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/openapi.json":{"get":{"operationId":"OpenApiController_getOpenApi","parameters":[],"responses":{"200":{"description":""}},"tags":["OpenApi"]}}},"info":{"title":"Nest-Next-OpenAPI-Template API","description":"The Nest-Next-OpenAPI-Template API documentation","version":"1.0","contact":{}},"tags":[{"name":"auth","description":"Authentication endpoints"},{"name":"users","description":"User management"}],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"},"cookie":{"type":"apiKey","in":"cookie","name":"access_token","description":"Access token stored in HTTP-only cookie"}},"schemas":{"SchemaHolderClass":{"type":"object","properties":{"detail":{"type":"string"}},"required":["detail"]}}}}